requirements.txt
alembic==1.13.3
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
asgiref==3.8.1
asyncpg==0.29.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==70.3.0
SQLAlchemy==2.0.29
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
